requirements.txt
streamlit==1.32.0
requests==2.31.0
tensorflow==2.13.0
pillow==10.0.0
pandas==2.2.0
numpy==1.24.0
opencv-python==4.8.0
plotly==5.18.0  # Optional for charts